chore: adding coverage ignore and moving mocks to test folder
