+++ src/rules/maxStatementsWithExclusions/maxStatementsWithExclusions.test.js
@@ -1,6 +1,6 @@
-const { RuleTester } = require("eslint");
-const rule = require("./index");
-const mockCallback = require("./mockCallback");
+const { RuleTester } = require('eslint');
+const rule = require('./index');
+const mockCallbackStatementExclusions = require('../../../test/unit/utils/mockCallbackStatementExclusionsStatementExclusions');
 
 const ruleTester = new RuleTester({ parserOptions: { ecmaVersion: 2020 } });
 
@@ -69,10 +69,10 @@ const invalidWithExclusions = [
 
 const optionsWithExclusions = [
   { max: 5 },
-  { exclusionCallback: require.resolve("./mockCallback") },
+  { exclusionCallback: require.resolve('./mockCallbackStatementExclusions') },
 ];
 
-ruleTester.run("max-statements-with-exclusions", rule, {
+ruleTester.run('max-statements-with-exclusions', rule, {
   valid: validWithExclusions.map((code) => ({
     code,
     options: optionsWithExclusions,
@@ -157,7 +157,17 @@ const optionsWithoutExclusion = [
 ];
 
 ruleTester.run(
-  "max-statements-excluding-specified-expression-type without exclusion",
+  'max-statements-excluding-specified-expression-type without exclusion',
+
+
+const optionsWithoutExclusion = [
+  {
+    max: 5,
+  },
+];
+
+ruleTester.run(
+  'max-statements-excluding-specified-expression-type without exclusion',
   rule,
   {
     valid: validWithoutExclusions.map((code) => ({
